fix : divide the serverManagerRequest into 2 files and fix comments
